Refactor: extract import-core.ts and break the import-modal/attachment-importer cycle
Move the 17 pure exports (error classification, formatting, result
tallying, paging merge, and the shared option/result types) out of
import-modal.ts into a new framework-free import-core.ts that imports no
obsidian API. import-modal.ts re-exports them so main.ts and the test
suite keep resolving unchanged; attachment-importer.ts now imports
ArtifactSelection from import-core, which removes the source-level cycle.

Add import-core.ts to the lint file list so the new module is covered by
the scorecard rules (the lint script uses an explicit file list).

No behavior change. tsc, lint, and 609 tests pass through the re-export.
